+++ qmate-package.json
@@ -3,7 +3,8 @@
     "@sap_oss/wdio-qmate-service": "^2.6.4",
     "chai": "^4.3.8",
     "install": "^0.13.0",
-    "npm": "^9.8.1"
+    "npm": "^9.8.1",
+    "wdio-qmate-reporter": "^1.5.0"
   },
   "devDependencies": {
     "@cucumber/cucumber": "^9.5.1",
@@ -18,6 +19,7 @@
   },
   "scripts": {
     "wdio": "wdio run ./config.js",
+    "wdio_cu": "wdio run ./config_cu.js",
     "allure": "allure generate ./allure-results --clean && allure open"
   }
 }
